+++ custom_components/bambu_lab/coordinator.py
@@ -28,13 +28,13 @@ from .pybambu.const import Features
 class BambuDataUpdateCoordinator(DataUpdateCoordinator):
     hass: HomeAssistant
     _updatedDevice: bool
-    _entry: ConfigEntry
+    latest_usage_hours: float
 
     def __init__(self, hass, *, entry: ConfigEntry) -> None:
         self._hass = hass
-        self._entry = entry
         LOGGER.debug(f"ConfigEntry.Id: {entry.entry_id}")
 
+        self.latest_usage_hours = float(entry.options.get('usage_hours', 0))
         self.client = BambuClient(device_type = entry.data["device_type"],
                                   serial = entry.data["serial"],
                                   host = entry.options['host'],
@@ -43,7 +43,8 @@ class BambuDataUpdateCoordinator(DataUpdateCoordinator):
                                   email = entry.options.get('email', ''),
                                   username = entry.options['username'],
                                   auth_token = entry.options['auth_token'],
-                                  access_code = entry.options['access_code'])
+                                  access_code = entry.options['access_code'],
+                                  usage_hours = self.latest_usage_hours)
             
         self._updatedDevice = False
         self.data = self.get_model()
@@ -77,6 +78,18 @@ class BambuDataUpdateCoordinator(DataUpdateCoordinator):
             elif event == "event_printer_data_update":
                 self._update_data()
 
+                # Check is usage hours change and persist to config entry if it did.
+                if self.latest_usage_hours != self.get_model().info.usage_hours:
+                    self.latest_usage_hours = self.get_model().info.usage_hours
+                    LOGGER.debug(f"OVERWRITING USAGE_HOURS WITH : {self.latest_usage_hours}")
+                    options = dict(self.config_entry.options)
+                    options['usage_hours'] = self.latest_usage_hours
+                    self._hass.config_entries.async_update_entry(
+                        entry=self.config_entry,
+                        title=self.get_model().info.serial,
+                        data=self.config_entry.data,
+                        options=options)
+
             elif event == "event_hms_errors":
                 self._update_hms()
 
@@ -97,7 +110,6 @@ class BambuDataUpdateCoordinator(DataUpdateCoordinator):
 
             elif event == "event_printer_cover_image_update":
                 self._update_data()
-
 
         async def listen():
             self.client.connect(callback=event_handler)
@@ -170,10 +182,10 @@ class BambuDataUpdateCoordinator(DataUpdateCoordinator):
         for index in range (0, len(device.ams.data)):
             if device.ams.data[index] is not None:
                 LOGGER.debug(f"Initialize AMS {index+1}")
-                hadevice = dev_reg.async_get_or_create(config_entry_id=self._entry.entry_id,
+                hadevice = dev_reg.async_get_or_create(config_entry_id=self.config_entry.entry_id,
                                                     identifiers={(DOMAIN, device.ams.data[index].serial)})
-                serial = self._entry.data["serial"]
-                device_type = self._entry.data["device_type"]
+                serial = self.config_entry.data["serial"]
+                device_type = self.config_entry.data["device_type"]
                 dev_reg.async_update_device(hadevice.id,
                                             name=f"{device_type}_{serial}_AMS_{index+1}",
                                             model="AMS",
@@ -185,10 +197,10 @@ class BambuDataUpdateCoordinator(DataUpdateCoordinator):
 
     def _update_external_spool_info(self):
         dev_reg = device_registry.async_get(self._hass)
-        hadevice = dev_reg.async_get_or_create(config_entry_id=self._entry.entry_id,
+        hadevice = dev_reg.async_get_or_create(config_entry_id=self.config_entry.entry_id,
                                                identifiers={(DOMAIN, f"{self.get_model().info.serial}_ExternalSpool")})
-        serial = self._entry.data["serial"]
-        device_type = self._entry.data["device_type"]
+        serial = self.config_entry.data["serial"]
+        device_type = self.config_entry.data["device_type"]
         dev_reg.async_update_device(hadevice.id,
                                     name=f"{device_type}_{serial}_ExternalSpool",
                                     model="External Spool",
@@ -214,8 +226,8 @@ class BambuDataUpdateCoordinator(DataUpdateCoordinator):
         return self.client.get_device()
 
     def get_printer_device(self):
-        printer_serial = self._entry.data["serial"]
-        device_type = self._entry.data["device_type"]
+        printer_serial = self.config_entry.data["serial"]
+        device_type = self.config_entry.data["device_type"]
 
         return DeviceInfo(
             identifiers={(DOMAIN, printer_serial)},
@@ -227,8 +239,8 @@ class BambuDataUpdateCoordinator(DataUpdateCoordinator):
         )
 
     def get_ams_device(self, index):
-        printer_serial = self._entry.data["serial"]
-        device_type = self._entry.data["device_type"]
+        printer_serial = self.config_entry.data["serial"]
+        device_type = self.config_entry.data["device_type"]
         device_name=f"{device_type}_{printer_serial}_AMS_{index+1}"
 
         return DeviceInfo(
@@ -242,8 +254,8 @@ class BambuDataUpdateCoordinator(DataUpdateCoordinator):
         )
 
     def get_virtual_tray_device(self):
-        printer_serial = self._entry.data["serial"]
-        device_type = self._entry.data["device_type"]
+        printer_serial = self.config_entry.data["serial"]
+        device_type = self.config_entry.data["device_type"]
         device_name=f"{device_type}_{printer_serial}_ExternalSpool"
 
         return DeviceInfo(